requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7  # PostgreSQL driver
Pillow==10.1.0
uuid
gunicorn
